+++ packages/config/src/api/site_info.js
@@ -1,4 +1,6 @@
+import { getEnvelope } from '../env/envelope.js'
 import { throwUserError } from '../error.js'
+import { ERROR_CALL_TO_ACTION } from '../log/messages.js'
 
 // Retrieve Netlify Site information, if available.
 // Used to retrieve local build environment variables and UI build settings.
@@ -7,6 +9,7 @@ import { throwUserError } from '../error.js'
 // Requires knowing the `siteId` and having the access `token`.
 // Silently ignore API errors. For example the network connection might be down,
 // but local builds should still work regardless.
+// eslint-disable-next-line complexity
 export const getSiteInfo = async function ({ api, siteId, mode, testOpts: { env: testEnv = true } = {} }) {
   if (api === undefined || mode === 'buildbot' || !testEnv) {
     const siteInfo = siteId === undefined ? {} : { id: siteId }
@@ -18,6 +21,13 @@ export const getSiteInfo = async function ({ api, siteId, mode, testOpts: { env:
     getAccounts(api),
     getAddons(api, siteId),
   ])
+
+  if (siteInfo.use_envelope) {
+    const envelope = await getEnvelope({ api, accountId: siteInfo.account_slug, siteId })
+    // eslint-disable-next-line fp/no-mutation
+    siteInfo.build_settings.env = envelope
+  }
+
   return { siteInfo, accounts, addons }
 }
 
@@ -55,5 +65,3 @@ const getAddons = async function (api, siteId) {
     throwUserError(`Failed retrieving addons for site ${siteId}: ${error.message}. ${ERROR_CALL_TO_ACTION}`)
   }
 }
-
-const ERROR_CALL_TO_ACTION = `Double-check your login status with 'netlify status' or contact support with details of your error.`